style: move routes+handlers into api folder. Add dist to .gitignore
