+++ backend/.gitignore
@@ -1,3 +1,3 @@
 .vercel
-
+dist
 drizzle
