+++ backend/src/tests/auth.test.ts
@@ -1,4 +1,4 @@
-import { logIn, signUp } from "../auth/auth.services.js";
+import { logIn, signUp } from "../api/auth/auth.services.js";
 import { conn } from "../db/database.js";
 
 // Test Suite
